Clean up tuple caching
