+++ packages/arb-avm-cpp/avm_values/include/avm_values/pool.hpp
@@ -17,6 +17,7 @@
 #ifndef pool_hpp
 #define pool_hpp
 
+#include <avm_values/tuplestub.hpp>
 #include <avm_values/value.hpp>
 
 #include <array>
@@ -24,9 +25,11 @@
 #include <vector>
 
 struct RawTuple {
-    uint256_t cachedHash = 0;
+    HashPreImage cachedPreImage;
     std::vector<value> data;
     bool deferredHashing = true;
+
+    RawTuple() : cachedPreImage({}, 0), deferredHashing(true) {}
 };
 
 class TuplePool {
