+++ packages/arb-avm-cpp/avm_values/include/avm_values/value.hpp
@@ -19,13 +19,9 @@
 
 #include <avm_values/bigint.hpp>
 #include <avm_values/opcodes.hpp>
+#include <avm_values/valuetype.hpp>
 
 #include <nonstd/variant.hpp>
-
-// Proof values will only include values types up to TUPLE + 8 (11)
-// All types declared with types creater than 11 are used for the internal
-// marshalling format Used to pass values between the AVM and the validator
-enum ValueTypes { NUM, CODEPT, HASH_PRE_IMAGE, TUPLE, CODE_POINT_STUB = 12 };
 
 class TuplePool;
 class Tuple;
